Update README.md
TODO List (CRUD) is a simple project developed using JavaScript, CSS, 
and HTML. The user can add the number of to-do work details and you 
can see the details stored in the list form. Also, you can even update the 
details of the record if you want and delete item to remove it.